Import isEnsDomain from types
